+++ core/utils/balls_and_bins.h
@@ -9,8 +9,8 @@
 // Enumerates all possible groupings of identical objects, i.e. balls in bins.
 // Note: This happens to be the same problem as computing the exponents of the
 // veronese map.
-IndexMat2DArray balls_and_bins(const unsigned int ballCount,
-                               const unsigned int binCount,
+IndexMat2DArray balls_and_bins(const size_t ballCount,
+                               const size_t binCount,
                                const bool all = 0);
 
 #endif  // CORE_UTILS_BALLS_AND_BINS_H_
